Put the double-precision trig where the orientation path actually goes
The previous commit routed WWMath's sin/cos/tan through double. It changed
nothing: replaying the same two PC recordings on the resulting build produced
byte-identical checksums, FBCBD397 and C6565527, exactly as before. Reverted,
because an unproven change with a cost only muddies the next measurement.

It was aimed at the wrong functions. Thing::setOrientation -- which builds
m_transform, the matrix Object::crc hashes raw -- does not call WWMath::Cos
and WWMath::Sin. It calls the engine's own Cos and Sin from Lib/trig.h,
implemented in GameEngine/Source/Common/System/Trig.cpp, and those called
sinf/cosf directly.

That file already carried a comment reasoning about this exact problem: these
feed the hashed transform, so they must match the client, therefore call the
single-precision CRT entry points the client calls. The reasoning was right and
the conclusion inverted. The source line is not what decides the answer; the C
runtime behind it is. The client is VC6 for 32-bit x86, whose CRT has no real
single-precision transcendentals -- sinf promotes to double and evaluates on
x87. bionic has real ones, with their own error.

Measured against the client's contract, x87 with the control word setFPMode()
installs (PC=24, round-to-nearest), over 62801 angles spanning [-pi, pi]:

    x87 fsin  vs  sinf                      776 differ  (1.236%)
    x87 fcos  vs  cosf                      768 differ  (1.223%)
    x87 fsin  vs  (float)sin((double)x)       0 differ
    x87 fcos  vs  (float)cos((double)x)       0 differ

One angle in eighty, on the rotation of every moving object every frame. The
MSVC/x86 arm keeps calling the CRT, because there it already is this.

Only the Zero Hour tree: the Generals tree routes these through
WWMath::SinTrig instead and is not built for Android, so it is left alone.

+++ Core/Libraries/Source/WWVegas/WWMath/matrix3.h
@@ -361,7 +361,7 @@ WWINLINE Matrix3x3::Matrix3x3(const Vector3 & axis,float s_angle,float c_angle)
 
 WWINLINE void Matrix3x3::Set(const Vector3 & axis,float angle)
 {
-	Set(axis,(float)sin((double)(angle)),(float)cos((double)(angle)));
+	Set(axis,sinf(angle),cosf(angle));
 }
 
 WWINLINE void Matrix3x3::Set(const Vector3 & axis,float s,float c)
@@ -775,7 +775,7 @@ WWINLINE int operator != (const Matrix3x3 & a, const Matrix3x3 & b)
  *=============================================================================================*/
 WWINLINE void Matrix3x3::Rotate_X(float theta)
 {
-	Rotate_X((float)sin((double)(theta)),(float)cos((double)(theta)));
+	Rotate_X(sinf(theta),cosf(theta));
 }
 
 WWINLINE void Matrix3x3::Rotate_X(float s,float c)
@@ -809,7 +809,7 @@ WWINLINE void Matrix3x3::Rotate_X(float s,float c)
  *=============================================================================================*/
 WWINLINE void Matrix3x3::Rotate_Y(float theta)
 {
-	Rotate_Y((float)sin((double)(theta)),(float)cos((double)(theta)));
+	Rotate_Y(sinf(theta),cosf(theta));
 }
 
 WWINLINE void Matrix3x3::Rotate_Y(float s,float c)
@@ -844,7 +844,7 @@ WWINLINE void Matrix3x3::Rotate_Y(float s,float c)
  *=============================================================================================*/
 WWINLINE void Matrix3x3::Rotate_Z(float theta)
 {
-	Rotate_Z((float)sin((double)(theta)),(float)cos((double)(theta)));
+	Rotate_Z(sinf(theta),cosf(theta));
 }
 
 WWINLINE void Matrix3x3::Rotate_Z(float s,float c)
@@ -898,7 +898,7 @@ WWINLINE Matrix3x3 Create_X_Rotation_Matrix3(float s,float c)
 
 WWINLINE Matrix3x3 Create_X_Rotation_Matrix3(float rad)
 {
-	return Create_X_Rotation_Matrix3((float)sin((double)(rad)),(float)cos((double)(rad)));
+	return Create_X_Rotation_Matrix3(sinf(rad),cosf(rad));
 }
 
 /***********************************************************************************************
@@ -934,7 +934,7 @@ WWINLINE Matrix3x3 Create_Y_Rotation_Matrix3(float s,float c)
 
 WWINLINE Matrix3x3 Create_Y_Rotation_Matrix3(float rad)
 {
-	return Create_Y_Rotation_Matrix3((float)sin((double)(rad)),(float)cos((double)(rad)));
+	return Create_Y_Rotation_Matrix3(sinf(rad),cosf(rad));
 }
 
 /***********************************************************************************************
@@ -970,7 +970,7 @@ WWINLINE Matrix3x3 Create_Z_Rotation_Matrix3(float s,float c)
 
 WWINLINE Matrix3x3 Create_Z_Rotation_Matrix3(float rad)
 {
-	return Create_Z_Rotation_Matrix3((float)sin((double)(rad)),(float)cos((double)(rad)));
+	return Create_Z_Rotation_Matrix3(sinf(rad),cosf(rad));
 }
 
 WWINLINE void Matrix3x3::Rotate_Vector(const Matrix3x3 & A,const Vector3 & in,Vector3 * out)
